+++ mohaeng/src/main/java/com/mohaeng/club/club/application/service/command/ChangeClubRoleName.java
@@ -1,6 +1,6 @@
-package com.mohaeng.club.club.application.service;
+package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.ChangeParticipantRoleUseCase;
+import com.mohaeng.club.club.application.usecase.ChangeClubRoleNameUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
 import com.mohaeng.club.club.exception.ClubException;
@@ -11,17 +11,17 @@ import static com.mohaeng.club.club.exception.ClubExceptionType.NOT_FOUND_CLUB;
 
 @Service
 @Transactional
-public class ChangeParticipantRole implements ChangeParticipantRoleUseCase {
+public class ChangeClubRoleName implements ChangeClubRoleNameUseCase {
 
     private final ClubRepository clubRepository;
 
-    public ChangeParticipantRole(final ClubRepository clubRepository) {
+    public ChangeClubRoleName(final ClubRepository clubRepository) {
         this.clubRepository = clubRepository;
     }
 
     @Override
     public void command(final Command command) {
         Club club = clubRepository.findById(command.clubId()).orElseThrow(() -> new ClubException(NOT_FOUND_CLUB));
-        club.changeParticipantRole(command.memberId(), command.targetParticipantId(), command.clubRoleId());
+        club.changeRoleName(command.memberId(), command.clubRoleId(), command.roleName());
     }
 }
