+++ mohaeng/src/main/java/com/mohaeng/club/club/application/service/command/CreateClub.java
@@ -1,4 +1,4 @@
-package com.mohaeng.club.club.application.service;
+package com.mohaeng.club.club.application.service.command;
 
 import com.mohaeng.club.club.application.usecase.CreateClubUseCase;
 import com.mohaeng.club.club.domain.model.Club;
