+++ mohaeng/src/main/java/com/mohaeng/club/club/application/service/command/DelegatePresident.java
@@ -1,6 +1,6 @@
-package com.mohaeng.club.club.application.service;
+package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.ChangeDefaultRoleUseCase;
+import com.mohaeng.club.club.application.usecase.DelegatePresidentUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
 import com.mohaeng.club.club.exception.ClubException;
@@ -11,17 +11,17 @@ import static com.mohaeng.club.club.exception.ClubExceptionType.NOT_FOUND_CLUB;
 
 @Service
 @Transactional
-public class ChangeDefaultRole implements ChangeDefaultRoleUseCase {
+public class DelegatePresident implements DelegatePresidentUseCase {
 
     private final ClubRepository clubRepository;
 
-    public ChangeDefaultRole(final ClubRepository clubRepository) {
+    public DelegatePresident(final ClubRepository clubRepository) {
         this.clubRepository = clubRepository;
     }
 
     @Override
     public void command(final Command command) {
         Club club = clubRepository.findById(command.clubId()).orElseThrow(() -> new ClubException(NOT_FOUND_CLUB));
-        club.changeDefaultRole(command.memberId(), command.clubRoleId());
+        club.delegatePresident(command.memberId(), command.presidentCandidateParticipantId());
     }
 }
