+++ mohaeng/src/test/java/com/mohaeng/club/club/application/service/command/CreateClubRoleTest.java
@@ -1,4 +1,4 @@
-package com.mohaeng.club.club.application.service;
+package com.mohaeng.club.club.application.service.command;
 
 import com.mohaeng.club.club.application.usecase.CreateClubRoleUseCase;
 import com.mohaeng.club.club.domain.model.Club;
